pyproject.toml
[tool.poetry]
name = "wth-baymax"
version = "0.1.0"
description = "template for python api server"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "<3.13,>=3.10"
python-dotenv = "^1.0.1"
fastapi-slim = "^0.115.6"
uvicorn = "^0.34.0"
uvloop = "^0.21.0"
httptools = "^0.6.4"
transformers = "^4.47.1"
torch = "^2.5.1"
librosa = "^0.10.2.post1"
accelerate = "^1.2.1"
python-multipart = "^0.0.20"
soundfile = "^0.12.1"
xxhash = "^3.5.0"
bitsandbytes = "^0.45.0"
optimum = "^1.23.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
poethepoet = "^0.31.1"
rich = "^13.9.4"

[tool.poetry.group.jupyter]
optional = true

[tool.poetry.group.jupyter.dependencies]
ipykernel = "*"
ipympl = "*"
ipywidgets = "*"
ipywebrtc = "*"

[tool.poe.tasks]
dev = { cmd = "python dev.py", use_exec = true }
prod = "docker run --rm --gpus all -p 3000:3000 wth-baymax:latest"
test = "echo Not Implemented Yet!"
publish = "echo Not Implemented Yet!"
requirements = "poetry export -o requirements.txt --without-hashes"

[tool.poe.tasks.build]
cmd = """docker build -f Dockerfile . \
  -t wth-baymax:latest \
  -t wth-baymax:${tag} \
"""
deps = ["requirements"]
args = [{ name = "tag", positional = true, required = true }]

[tool.ruff.lint]
select = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
